+++ pumpkindb_mio_server/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "pumpkindb_server"
+name = "pumpkindb_mio_server"
 version = "0.2.0"
 license = "MPL-2.0"
 repository = "https://github.com/PumpkinDB/PumpkinDB"
@@ -9,20 +9,17 @@ categories = [ "database" ]
 
 authors = ["Yurii Rashkovskii <yrashk@gmail.com>"]
 
-[[bin]]
-doc = false
-name = "pumpkindb"
-
 [dependencies]
-clap = "2.22.1"
-config = "0.3.1"
-lazy_static = "0.2.2"
-num_cpus = "1.3.0"
+mio = "0.6.4"
+byteorder = "1.0.0"
 memmap = "0.5.2"
+slab = "0.3.0"
+num-bigint = "0.1.35"
+num-traits = "0.1.36"
+rand = "0.3.15"
 log = "0.3.6"
 log4rs = { version = "0.6.1", features = ["toml_format"] }
 lmdb-zero = "0.4.0"
-mio = "0.6.4"
+uuid = { version = "0.4", features = ["v4"] }
 
 pumpkindb_engine = { version = "0.2", path = "../pumpkindb_engine" }
-pumpkindb_mio_server = { version = "0.2", path = "../pumpkindb_mio_server" }
